Move `CoinPerByte` to `cardano-ledger-core`
